chore: pull latest helio
Also clean-up unused flags.
